Replace scrot with maim
Scrot has several issues such as poor line rendering when using mouse
selection where lines were being drawn all over the selected screenshot.
Additionally, the screenshot name has been updated to record the time of
the actual screenshot rather than when the script was invoked. For
example, when mouse selection is used, this is the time the mouse is released
and the actual screenshot is taken.