Cannot use $TERM for terminal name
Recently a change was made to rename the terminal from "st-256color" to "st".
This allows calling the $TERM variable to launch the terminal program st.
However, this is not reliable as the $TERM variable is not fixed.
The variable is set and reset in different places, and is not
consistently "st". For example, it may be changed to "linux".
Thus, keep the $TERMINAL variable to invoke the terminal client.