Check the modification time of all subfiles
The modification time of a directory is not recursive. There can be a
subfile within the directory that has a newer modification time than the
parent directory. Thus, we also need to check if any subfile of a
direcotry is newer than the archive to determine if it should be
re-archived.