Kill pipewire upon exiting Hyprland
Pipewire persists in the background even when exiting Hyprland and even
when logging out of the tty. This causes issues when launching Hyprland
again as the existing Pipewire instance isn't configured to the current
instance and no sound is produced. This fix allows a new instance of
Pipewire to be launched every time.